+++ NEWS.md
@@ -5,6 +5,7 @@ NEWS
 
 # TOSTER v0.8.6
 - Add warning message about error control with MET 
+- Fix unit tests for boot_ses_calc to catch errors when estimates contain infinite values or when ses is not "rb"
 
 # TOSTER v0.8.5
 - Big update to package documentation to make things more detailed.